requirements.txt
skytemple-files==0.1.0a15
Pillow==7.1.2
